CAS-364: Refactor server code to reduce duplicated functionality
