+++ backend/.env.example
@@ -1,3 +1,4 @@
+API_PORT="5001"
 DB_USERNAME="postgres"
 DB_PASSWORD="admin"
 DB_NAME="flow_snapshot"
